requirements.txt
beautifulsoup4==4.9.0
blinker==1.4
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.1
Flask==1.1.2
Flask-Mail==0.9.1
gunicorn==20.0.4
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.14.1
MarkupSafe==1.1.1
nltk==3.5
numpy==1.18.3
PyJWT==1.7.1
PySocks==1.7.1
python-Levenshtein==0.12.0
pytz==2019.3
regex==2020.4.4
requests==2.23.0
scikit-learn==0.22.2.post1
scipy==1.4.1
six==1.14.0
sklearn==0.0
soupsieve==2.0
tqdm==4.45.0
twilio==6.0.0
urllib3==1.25.9
Werkzeug==1.0.1
